Release 4 (#945)
* add question_uuid option for function_completion

* add ability to pass text to help (#813)

* lets just say deleted

* say deleted

* onward

* next

* wtf

* doh

* Save message key to science server (#814)

* Save message on cache-manager and send message id to science server.

* Fix key label.

* Err level log instead of debug.

* Add ttl.

* Revert "Add ttl."

This reverts commit 8b95fc0d9fa0100daa7477ae4bbf8600a4e543c0.

* Fix messsage id ref.

* Revert "Revert "Add ttl.""

This reverts commit 692dc3e1f29d3e74c4cf785dcbcb04ad0130891b.

* Revert "Revert "Revert "Add ttl."""

This reverts commit d253f60f6557cab88e65aca69763721e05464b77.

---------


* next

* tweak

* fix type

* fix tests by mocking redis

* set default for localhost

* use host url everywhere!

* onward

* Remember tree state. (#823)


* make contactEmail and legalUrl required fields!

* Usage of Promise.all to reduce response time. (#806)


* switch to token_set_ratio

* NPM package build, deploy and tag automation (#826)

* test new pipe cicd

* NPM package CI/CD

* Cicd improvement  (#829)

* test new pipe cicd

* NPM package CI/CD

* added validation of npm package version

* test new pipe CI/CD

* last changes on pipe CI/CD

* Cicd improvement  (#833)

* test new pipe cicd

* NPM package CI/CD

* added validation of npm package version

* Test Pipe CI/CD

* upgrades to pipes CI/CD

* fix identation error

* test

* Cicd improvement fm (#834)

* test new pipe cicd

* NPM package CI/CD

* added validation of npm package version

* Test Pipe CI/CD

* upgrades to pipes CI/CD

* fix identation error

* Cicd improvement fm (#835)

* test new pipe cicd

* NPM package CI/CD

* added validation of npm package version

* Test Pipe CI/CD

* upgrades to pipes CI/CD

* fix identation error

* Cicd improvement fm (#836)

* test new pipe cicd

* Cicd improvement fm (#838)

* test new pipe cicd

* final changes

* test pipe CI/CD

* test pipe

* test

* Fixes. (#853)


* reset registry of development dockerimage

* fix

* nested args (#862)

* next

* progress

* sometimes ya gotta brute force it

* next

* fix tests

* remove chat

* up the limit for desc for model

* Arrange some stuff (#863)

* Avoid double `json.loads` and rebuilt extension.

* Handle parsing err on message event.

---------


* Fix optional non string args (#869)

* Remove keys from function arguments that are not specified on execution call.

* Better naming.

* Typo on comment.

* Added new log.

* Parsed arguments meta data.

---------


* add workspaceFolder support (#828)

* onward

* onward

* next

* Added `IsString()`.

* Pass workspace folder from vscode extension to server.

* Load conversation history based on workspace folder.

* Fix saving workspace folder.

* Upgraded version on package.json

---------


* Added hash on string args too, to remove them if not present.

* fix infinite conversation history bug

* actually the base gpt type is fine

* Fallback to empty array.

* Serverside limits on poly free account (#68) (#875)

* Plugin Conversation Lookback (#874)

* almost

* onward

* next

* next

* next

* fix everything

* lets see whole conversation for gptplugin

* lets fix the bug

* making things better

* better strip type

* remove namby -pamby, wishy-washy language from description prompt

* better

* Serverside limits enhancements (#68) (#882)

* Variable names are not being validated (#891) (#892)

* better keyword translate to english

* added pstgresql dependency

* little fix

* little fix

* Err handling on `message` event for streaming api. (#871)

* Err handling on `message` event for streaming api.

* Fix optional non string args (#869)

* Remove keys from function arguments that are not specified on execution call.

* Better naming.

* Typo on comment.

* Added new log.

* Parsed arguments meta data.

---------


* Added hash on string args too, to remove them if not present.

---------


* set convo id (#900)

* killing it

* fix tests

* next

* onward to proper json messages

* better

* next

* wooot are we there?

* lock down so you can only add to your own conversation

* An endpoint to see public functions/variables and control their visibility within my tenant (#709) (#893)

* Tenant sign up (#899)

* Advances.

* Advances.

* More err handling.

* Email validation in client side.

* Added availability api.

* Better flow.

* Email sending.

* Send email with details.

* Last adjustments.

* Added tos logic.

* Log response in provider.

* Some minor fixes.

* Refactor.

* Bugfix.

* Last changes.

* Simplify PR.

* Simplify pr #2.

* Added mocks.

* Fix url.

* Changed tos to tos-agreements on tenant routes.

* Assign tos automatically.

* Corrections #1.

* Corrections #2 - Ran lint inside @poly/model.

---------


* Mount tenant's environment on tenant's creation.

* Rebuilt extension.

* Enforce HTTPS in Polyapi client library (#883, #884) (#894)

* Sign up adjustments.

* Upgrade patch version.

* fix message type

* Ability to specify a default Tier for new tenant creation (#897) (#904)

* Fix vscode crendentials sync (#908)

* Fixes.

* Use fspath instead of path.

* Done fixes.

---------


* Last changes on tenant sign up flow.

* Tenant sign up name optional in database.

* Override arrays with empty ones. (#911)


* Correction in question message for tos.

* Update a new question.

* Return server response to webhook caller (#809) (#903)

* Little fix

* Support optional elements inside an argument object (#898) (#912)

* Trigger name attribute not being respected in the POST call (#850) (#918)

* Sign up email templates (#919)

* Templates usage.

* Removed duplicated log.

---------


* add variable match limit (#924)

* OpenAPI spec for consumers, where super admin is not included (#878) (#920)

* Permission for managing triggers (#842) (#922)

* allow-application-conversations (#909)

* onward

* maybe this is it

* next

* fix store_messages

* getting there

* onward

* fix tsts

* next

* killing it

* getting there

* better

* omit history if maximum content length too long

* allow ANYONE to list the plugins in their own environment

* Add variable limit to tier configuration and count variable CRUD as function calls (#905) (#925)

* Added waitForResponse to Trigger (#809) (#935)

* Clean up tenant discrepancies (#931) (#936)

* Ability to disable a custom function (#879) (#934)

* Trigger name should be unique (#842) (#937)

* by default show ALL workspaceFolder convos

* Allow users set an arbitrary instance when creating tenants.

* Removed IsNotEmpty constraint.

* Upgrade version.

* 711 default tos (#942)

* Configure default tos.

* Moved not found err logic to controller.

---------


---------

+++ src/trigger/trigger.controller.ts
@@ -1,9 +1,9 @@
-import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Req, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Logger, NotFoundException, Param, Post, Req, UseGuards } from '@nestjs/common';
 import { ApiSecurity } from '@nestjs/swagger';
 import { PolyAuthGuard } from 'auth/poly-auth-guard.service';
 import { AuthData, AuthRequest } from 'common/types';
 import { TriggerService } from 'trigger/trigger.service';
-import { CreateTriggerDto, Role, TriggerDestination, TriggerSource } from '@poly/model';
+import { CreateTriggerDto, Permission, TriggerDestination, TriggerDto, TriggerResponseDto, TriggerSource, Visibility } from '@poly/model';
 import { WebhookService } from 'webhook/webhook.service';
 import { FunctionService } from 'function/function.service';
 import { AuthService } from 'auth/auth.service';
@@ -11,6 +11,8 @@ import { AuthService } from 'auth/auth.service';
 @ApiSecurity('PolyApiKey')
 @Controller('triggers')
 export class TriggerController {
+  private readonly logger = new Logger(TriggerController.name);
+
   constructor(
     private readonly triggerService: TriggerService,
     private readonly webhookService: WebhookService,
@@ -19,33 +21,52 @@ export class TriggerController {
   ) {
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get()
   async getTriggers(@Req() req: AuthRequest) {
+    await this.authService.checkPermissions(req.user, Permission.ManageTriggers);
+
     return await this.triggerService.getTriggers(req.user.environment.id);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Post()
   async createTrigger(@Req() req: AuthRequest, @Body() data: CreateTriggerDto) {
-    const { source, destination } = data;
+    const { name = null, source, destination, waitForResponse = false } = data;
+
+    await this.authService.checkPermissions(req.user, Permission.ManageTriggers);
     await this.checkTriggerSource(req, source);
     await this.checkTriggerDestination(req, destination);
 
-    return await this.triggerService.createTrigger(req.user.environment.id, source, destination);
+    return await this.triggerService.createTrigger(req.user.environment.id, name, source, destination, waitForResponse);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get(':id')
   async getTrigger(@Req() req: AuthRequest, @Param('id') id: string) {
-    return await this.findTrigger(req.user, id);
+    const trigger = await this.findTrigger(req.user, id);
+
+    await this.checkTriggerAccess(req.user, trigger);
+
+    return trigger;
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Delete(':id')
   async deleteTrigger(@Req() req: AuthRequest, @Param('id') id: string) {
     const trigger = await this.findTrigger(req.user, id);
+
+    await this.checkTriggerAccess(req.user, trigger);
+
     return await this.triggerService.deleteTrigger(req.user.environment.id, trigger);
+  }
+
+  @Post('/response')
+  async postTriggerResponse(@Body() response: TriggerResponseDto) {
+    const { executionId, data } = response;
+    this.logger.debug(`Received trigger response ${executionId} with data ${JSON.stringify(data)}`);
+
+    await this.triggerService.processTriggerResponse(executionId, data);
   }
 
   private async findTrigger(authData: AuthData, id: string) {
@@ -82,5 +103,12 @@ export class TriggerController {
     if (!serverFunction || !await this.authService.hasEnvironmentEntityAccess(serverFunction, req.user, true)) {
       throw new NotFoundException(`Server function ${functionId} not found`);
     }
+  }
+
+  private async checkTriggerAccess(authData: AuthData, trigger: TriggerDto) {
+    await this.authService.checkEnvironmentEntityAccess({
+      ...trigger,
+      visibility: Visibility.Environment,
+    }, authData, false, Permission.ManageTriggers);
   }
 }
